Added delete option for Default tracks.

+++ src/com/agilecrm/core/api/deals/MilestoneAPI.java
@@ -10,7 +10,9 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 
 import com.agilecrm.deals.Milestone;
 import com.agilecrm.deals.util.MilestoneUtil;
@@ -98,6 +100,11 @@ public class MilestoneAPI
     @Produces({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
     public Milestone savePipeline(Milestone milestone)
     {
+	if (MilestoneUtil.isDuplicate(milestone, false))
+	{
+	    throw new WebApplicationException(Response.status(Response.Status.BAD_REQUEST)
+		    .entity("Sorry, Track already exists with this name.").build());
+	}
 	milestone.save();
 	return milestone;
     }
@@ -113,6 +120,17 @@ public class MilestoneAPI
     @Produces({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
     public Milestone updatePipeline(Milestone milestone)
     {
+	if (milestone.name.equals("Default") && MilestoneUtil.isDuplicate(milestone, true))
+	{
+	    throw new WebApplicationException(Response.status(Response.Status.BAD_REQUEST)
+		    .entity("Sorry, There should be atleast one Track with name Default.").build());
+	}
+
+	if (MilestoneUtil.isDuplicate(milestone, true))
+	{
+	    throw new WebApplicationException(Response.status(Response.Status.BAD_REQUEST)
+		    .entity("Sorry, Track already exists with this name.").build());
+	}
 	milestone.save();
 	return milestone;
     }
@@ -128,6 +146,19 @@ public class MilestoneAPI
     public Milestone deletePipeline(@PathParam("id") Long id)
     {
 	Milestone milestone = MilestoneUtil.getMilestone(id);
+	// Throw non-200 if it exists
+	if (milestone.name.equals("Default") && !MilestoneUtil.isDuplicate(milestone, true))
+	{
+	    throw new WebApplicationException(Response.status(Response.Status.BAD_REQUEST)
+		    .entity("Sorry, There should be atleast one Default Track.").build());
+	}
+
+	List<Milestone> milestones = MilestoneUtil.getMilestonesList();
+	if (milestones.size() == 1)
+	{
+	    throw new WebApplicationException(Response.status(Response.Status.BAD_REQUEST)
+		    .entity("Sorry, There should be atleast one track.").build());
+	}
 	milestone.delete();
 	return milestone;
     }
